Consolidate usages of the _transferer hack.
